Safari-611.1.10/LayoutTests/media/media-controller-unpause-expected.txt


RUN(controller = video.controller)
EVENT(canplaythrough)
RUN(controller.pause())
EVENT(pause)
RUN(controller.unpause())
EVENT(play)
EXPECTED (controller.playbackState == 'waiting') OK
RUN(controller.play())
EVENT(playing)
END OF TEST
